requirements.txt
fastapi==0.136.1
uvicorn[standard]==0.34.0
pydantic==2.13.3
pydantic-settings==2.9.1
sqlalchemy==2.0.41
alembic==1.16.1
python-jose[cryptography]==3.5.0
passlib[bcrypt]==1.7.4
bcrypt==4.3.0
slowapi==0.1.9
APScheduler==3.11.0
python-multipart==0.0.20
email-validator==2.3.0
python-dotenv==1.1.0
sqladmin==0.20.1
itsdangerous==2.2.0
